hunting the manticore
"Boss battle" meant to get me used to the idea of using methods. Took me a few hours

It works, but i cant figure out how to get a win message to display. The while loop breaks when manticoreHealth < 0 as is expected but it wont display anything else. I'm going to revisit this when I have time to watch the Author's video on youtube.